Unify ConfigurationManager and ModuleConfiguration
