+++ CMakeLists.txt
@@ -33,7 +33,7 @@ daq_protobuf_codegen( opmon/*.proto )
 
 ##############################################################################
 # Main library
-daq_add_library(Application.cpp DAQModule.cpp DAQModuleManager.cpp ConfigurationManager.cpp ModuleConfiguration.cpp
+daq_add_library(Application.cpp DAQModule.cpp DAQModuleManager.cpp ConfigurationManager.cpp
   LINK_LIBRARIES ${APPFWK_DEPENDENCIES})
 
 ##############################################################################
