+++ src/DAQModuleManager.cpp
@@ -42,15 +42,15 @@ DAQModuleManager::DAQModuleManager()
 void
 DAQModuleManager::initialize(std::shared_ptr<ConfigurationManager> cfgMgr, opmonlib::OpMonManager& opm)
 {
-  m_module_configuration = std::make_shared<ModuleConfiguration>(cfgMgr);
-  get_iomanager()->configure(cfgMgr->session()->UID(),
-                             m_module_configuration->queues(),
-                             m_module_configuration->networkconnections(),
-                             m_module_configuration->connectivity_service(),
+  m_configuration_mgr = cfgMgr; // Make a copy
+  get_iomanager()->configure(m_configuration_mgr->session()->UID(),
+                             m_configuration_mgr->queues(),
+                             m_configuration_mgr->networkconnections(),
+                             m_configuration_mgr->connectivity_service(),
                              opm);
-  init_modules(m_module_configuration->modules(), opm);
+  init_modules(m_configuration_mgr->modules(), opm);
 
-  for (auto& plan_pair : m_module_configuration->action_plans()) {
+  for (auto& plan_pair : m_configuration_mgr->action_plans()) {
     auto cmd = plan_pair.first;
 
     for (auto& step : plan_pair.second->get_steps()) {
@@ -297,7 +297,7 @@ DAQModuleManager::execute(const std::string& cmd, const dataobj_t& cmd_data)
 
   check_cmd_data(cmd, cmd_data);
 
-  auto action_plan = m_module_configuration->action_plan(cmd);
+  auto action_plan = m_configuration_mgr->action_plan(cmd);
   if (action_plan == nullptr) {
 #if 0
     throw ActionPlanNotFound(ERS_HERE, cmd, "Throwing exception");
