+++ src/DAQModuleManager.hpp
@@ -13,7 +13,6 @@
 #include "nlohmann/json.hpp"
 
 #include "appfwk/ConfigurationManager.hpp"
-#include "appfwk/ModuleConfiguration.hpp"
 #include "confmodel/DaqModule.hpp"
 #include "conffwk/Configuration.hpp"
 
@@ -99,7 +98,7 @@ private:
   void check_mod_has_cmd(const std::string& cmd, const std::string& mod_class, const std::string& mod_id = "");
 
   std::vector<std::string> get_modnames_by_cmdid(cmdlib::cmd::CmdId id);
-  std::shared_ptr<ModuleConfiguration> m_module_configuration;
+  std::shared_ptr<ConfigurationManager> m_configuration_mgr;
 
   bool m_initialized;
